# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/car_photo": {
            "get": {
                "tags": [
                    "Car_Photo"
                ],
                "summary": "Get All Car Photos",
                "operationId": "db80130f5bdef06073ab42a3fb74a098",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "car_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "image_url": {
                                                        "type": "string",
                                                        "example": "https://example.com/images/car1.jpg"
                                                    },
                                                    "is_main": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Car_Photo"
                ],
                "summary": "Create Car Photo",
                "operationId": "692bc4e83212537fe118b2ba6f632a72",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "car_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "image_url": {
                                                "type": "string",
                                                "example": "https://example.com/images/car1.jpg"
                                            },
                                            "is_main": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "car_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "image_url": {
                                                    "type": "string",
                                                    "example": "https://example.com/images/car1.jpg"
                                                },
                                                "is_main": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/car_photo/{car_photo}": {
            "get": {
                "tags": [
                    "Car_Photo"
                ],
                "summary": "Get Photo by param",
                "operationId": "af593b8145bea5e9d18ffe067a5b119b",
                "parameters": [
                    {
                        "name": "car_photo",
                        "in": "path",
                        "description": "PostId",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "car_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "image_url": {
                                                        "type": "string",
                                                        "example": "https://example.com/images/car1.jpg"
                                                    },
                                                    "is_main": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Car_Photo"
                ],
                "summary": "Delete Car_photo by id",
                "operationId": "89c86c3e57cf739d6fc08b6b5af6eb1e",
                "parameters": [
                    {
                        "name": "car_photo",
                        "in": "path",
                        "description": "PostId",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Deleted car photo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Car_Photo"
                ],
                "summary": "Update Car Photo by param",
                "operationId": "9f17e01101c7eaf98304dab896bcd0ab",
                "parameters": [
                    {
                        "name": "car_photo",
                        "in": "path",
                        "description": "PostId",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "car_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "image_url": {
                                                "type": "string",
                                                "example": "https://example.com/images/car1.jpg"
                                            },
                                            "is_main": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "car_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "image_url": {
                                                        "type": "string",
                                                        "example": "https://example.com/images/car1.jpg"
                                                    },
                                                    "is_main": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}